requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
pydantic-settings
google-generativeai
slowapi
pyjwt
passlib
python-multipart
azure-keyvault-secrets
azure-identity
python-dotenv
azure-storage-blob
httpx

# NLP and text processing
spacy>=3.8.0

# LangChain dependencies
langchain==0.3.27
langchain-core==0.3.79
langchain-openai==0.3.33
langchain-community==0.3.29
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.9


